Issue #31207979 by robertragas: add email fields to optional and add uninstall hook
