+++ modules/social_features/social_event/modules/social_event_an_enroll/config/optional/field.storage.event_enrollment.field_email.yml
@@ -3,9 +3,6 @@ status: true
 dependencies:
   module:
     - social_event
-  enforced:
-    module:
-      - social_event_an_enroll
 id: event_enrollment.field_email
 field_name: field_email
 entity_type: event_enrollment
