+++ modules/social_features/social_event/modules/social_event_invite/config/optional/field.field.event_enrollment.event_enrollment.field_email.yml
@@ -1,0 +1,19 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.event_enrollment.field_email
+  module:
+    - social_event
+id: event_enrollment.event_enrollment.field_email
+field_name: field_email
+entity_type: event_enrollment
+bundle: event_enrollment
+label: 'Email address'
+description: 'Enter your email, so we can send you event updates.'
+required: true
+translatable: false
+default_value: {  }
+default_value_callback: ''
+settings: {  }
+field_type: email
