+++ modules/social_features/social_event/modules/social_event_invite/config/optional/field.storage.event_enrollment.field_email.yml
@@ -1,0 +1,17 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - social_event
+id: event_enrollment.field_email
+field_name: field_email
+entity_type: event_enrollment
+type: email
+settings: {  }
+module: core
+locked: false
+cardinality: 1
+translatable: true
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
