+++ modules/social_features/social_event/modules/social_event_invite/social_event_invite.install
@@ -1,0 +1,23 @@
+<?php
+
+use Drupal\field\Entity\FieldConfig;
+
+/**
+ * @file
+ * The Social event invite enroll module.
+ */
+
+/**
+ * Implements hook_uninstall().
+ */
+function social_event_invite_uninstall() {
+  $moduleHandler = \Drupal::service('module_handler');
+  // Delete the email field if social_event_an_enroll is not enabled.
+  // Todo:: is there a way to get field dependencies instead of manual?
+  if (!$moduleHandler->moduleExists('social_event_an_enroll')){
+    $field = FieldConfig::loadByName('event_enrollment', 'event_enrollment', 'field_email');
+    if ($field) {
+      $field->delete();
+    }
+  }
+}
